Allow date of birth to be supplied as a Date or Time
Forcing a String parameter is a bit restrictive, particularly when using
a framework that automatically converts dates as they come out of
database (e.g. Rails / ActiveRecord).

The user shouldn't need to get involved in the implementation details
given how easy it is for the library to convert to a String